CMFileManager: Fix problem when "Use swipe gestures" on
When "Use swipe gestures" is on, click on listview item has no
effects.
Since ev.getX() or ev.getY() is returning floating point, while
the previous touch location is stored as fixed point. When
there is fractional parts in getX(), it will always return
false when it compares to the previous touch point, making the flinger
logic to think the touch does moved while it is actually not.

FEIJ-290

(cherry picked from commit fb5b91434a10910ce29395293f8430b0eb9a10c6)

+++ src/com/cyanogenmod/filemanager/ui/widgets/FlingerListView.java
@@ -375,7 +375,7 @@ public class FlingerListView extends ListView {
             if (this.mFlingingView != null) {
                 // Only if event has changed (and only to the right and if not scrolling)
                 if (!this.mScrolling) {
-                    if (ev.getX() >= this.mStartX && (ev.getX() - this.mCurrentX != 0)) {
+                    if ((int)ev.getX() >= this.mStartX && ((int)ev.getX() - this.mCurrentX != 0)) {
                         this.mCurrentX = (int)ev.getX();
                         this.mTranslationX = this.mCurrentX - this.mStartX;
                         this.mFlingingView.setTranslationX(this.mTranslationX);
